requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
transformers>=4.27.0
sentencepiece>=0.1.99
tokenizers>=0.13.0
torch>=1.13.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.64.0
requests>=2.28.0
more-itertools>=8.14.0
pyyaml>=6.0
paddlepaddle>=2.4.0,<3.0.0  # 需要 <3.0.0 版本以兼容 paddlenlp 模型转换
paddlenlp>=2.4.0
aistudio-sdk<0.3.0  # 需要 <0.3.0 版本以兼容 paddlenlp
opencc-python-reimplemented>=0.1.7
thefuzz>=0.19.0
wikipedia-api>=0.6.0
zhconv>=1.4.2
